Refactor each subcommand to its own file
No other changes, just moving code.
